backend/requirements.txt
Flask==3.0.0
firebase-admin==6.2.0
flask_cors==4.0.0
pandas==2.1.0
numpy==1.24.3
pickle-mixin==1.0.2
google-cloud-storage==2.12.0

